// test/integration/test/ports/prisma/functional/relation-mode-gh-m-to-n/_fixture/default-nomap/generated/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgAdapterQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly CategoriesOnPostsManyToMany: {
      readonly postId: CodecTypes['pg/text@1']['output'];
      readonly categoryId: CodecTypes['pg/text@1']['output'];
    };
    readonly CategoryManyToMany: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly published: CodecTypes['pg/bool@1']['output'] | null;
    };
    readonly PostManyToMany: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly published: CodecTypes['pg/bool@1']['output'] | null;
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly CategoriesOnPostsManyToMany: {
      readonly postId: CodecTypes['pg/text@1']['input'];
      readonly categoryId: CodecTypes['pg/text@1']['input'];
    };
    readonly CategoryManyToMany: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly published: CodecTypes['pg/bool@1']['input'] | null;
    };
    readonly PostManyToMany: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly published: CodecTypes['pg/bool@1']['input'] | null;
    };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly categoriesOnPostsManyToMany: {
      readonly categoryId: CodecTypes['pg/text@1']['output'];
      readonly postId: CodecTypes['pg/text@1']['output'];
    };
    readonly categoryManyToMany: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly published: CodecTypes['pg/bool@1']['output'] | null;
    };
    readonly postManyToMany: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly published: CodecTypes['pg/bool@1']['output'] | null;
    };
  };
};
export type StorageColumnInputTypes = {
  readonly public: {
    readonly categoriesOnPostsManyToMany: {
      readonly categoryId: CodecTypes['pg/text@1']['input'];
      readonly postId: CodecTypes['pg/text@1']['input'];
    };
    readonly categoryManyToMany: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly published: CodecTypes['pg/bool@1']['input'] | null;
    };
    readonly postManyToMany: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly published: CodecTypes['pg/bool@1']['input'] | null;
    };
  };
};
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;